+++ src/app/components/icon/index.ts
@@ -1,2 +1,2 @@
-export * from "./icon";
+export * from "./8con";
 export * from "./icon.type";
